feat: 유저별 학습 현황 조회 서비스 구현 & 테스트 중복 메서드 추출

+++ devly-domain/src/main/java/se/sowl/devlydomain/study/domain/StudyType.java
@@ -19,10 +19,10 @@ public class StudyType {
     private String name;
 
     @Column(name = "base_exp")
-    private Integer baseExp;
+    private Long baseExp;
 
     @Builder
-    public StudyType(String name, Integer baseExp) {
+    public StudyType(String name, Long baseExp) {
         this.name = name;
         this.baseExp = baseExp;
     }
